move cdn metadata model to the metadata pkg
