+++ pkg/cdn/metadata/cdn_metadata_entry.go
@@ -1,4 +1,4 @@
-package models
+package metadata
 
 import (
 	"encoding/json"
