+++ pkg/cdn/metadata/rdbms/getter.go
@@ -1,10 +1,10 @@
 package rdbms
 
-import "github.com/gomods/athens/pkg/cdn/metadata/rdbms/models"
+import "github.com/gomods/athens/pkg/cdn/metadata"
 
 // Get retrieves the cdn base URL for a module
 func (s *MetadataStore) Get(module string) (string, error) {
-	result := models.CdnMetadataEntry{}
+	result := metadata.CdnMetadataEntry{}
 	query := s.conn.Where("module = ?", module)
 	if err := query.First(&result); err != nil {
 		return "", err
